% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{casualties}
\alias{casualties}
\title{casualty data from the Afghan War Diary, lightly cleaned}
\format{\preformatted{'data.frame':	50414 obs. of  12 variables:
 $ date    : POSIXct, format: "2004-01-18 02:00:00" "2004-01-18 02:00:00" ...
 $ year    : num  2004 2004 2004 2004 2004 ...
 $ month   : Ord.factor w/ 12 levels "January"<"February"<..: 1 1 1 1 1 2 2 2 2 2 ...
 $ day     : Ord.factor w/ 7 levels "Sunday"<"Monday"<..: 1 1 1 2 2 6 6 6 3 3 ...
 $ lon     : num  65.5 65.5 65.5 71.1 71.1 ...
 $ lat     : num  32.6 32.6 32.6 34.8 34.8 ...
 $ fatal   : logi  TRUE TRUE TRUE FALSE FALSE FALSE ...
 $ harmful : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ type    : chr  "Enemy Action" "Enemy Action" "Enemy Action" "Explosive Hazard" ...
 $ category: chr  "Direct Fire" "Direct Fire" "Direct Fire" "IED Explosion" ...
 $ victim  : Factor w/ 4 levels "enemy","civilian",..: 4 4 4 4 4 4 4 4 4 4 ...
 $ casualty: chr  "wia" "wia" "wia" "wia" ...
}}
\usage{
casualties
}
\description{
The Afghan War Diary is distributed by the WikiLeaks organization and available
for download at http://www.wikileaks.org/wiki/Afghan_War_Diary,_2004-2010. Each
row of the data set corresponds to a single casualty
(one person being injured or killed).
}
\keyword{datasets}

