\name{condense}
\alias{condense}
\title{reduce a single valued vector to a single element}
\usage{
  condense(var)
}
\arguments{
  \item{var}{a vector}
}
\value{
  a vector of length one if var only contains one unique
  value, var otherwise
}
\description{
  reduce a single valued vector to a single element
}
\keyword{internal}

