\name{first_name}
\alias{first_name}
\title{Get the first name in an expression}
\usage{
  first_name(expr)
}
\arguments{
  \item{expr}{an \code{\link{expression}} or
  \code{\link{call}} from which names are to be extracted}
}
\value{
  character
}
\description{
  first_name retrieves the first name used in an
  expression. To be used with screening mappings for plyr
  or regular computation.
}
\keyword{internal}

