% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{get_xs}
\alias{get_xs}
\title{Get all mappings related to plotting on the x axis}
\usage{
get_xs(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
get_xs retrieves all mappings that need to be altered for plotting on a new x axis
}
\keyword{internal}

