% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggsubplot-class.r
\docType{class}
\name{ggsubplot-class}
\alias{ggsubplot-class}
\alias{show,ggsubplot-method}
\title{ggsubplot class}
\description{
a ggsubplot object is a ggplot object that has been extended to include methods
for embedding subplots when plotting.
}

