% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggsubplot-class.r
\name{ggsubplot}
\alias{ggsubplot}
\title{Create a ggsubplot object}
\usage{
ggsubplot(ggplot)
}
\arguments{
\item{ggplot}{a gg or ggplot object}
}
\description{
glyph_plot gives a ggplot object the S4 class `ggsubplot', see
\code{\link{ggsubplot-class}}. ggsubplot denotes ggplot objects that contain
extra information to be used to embed subplots when plotting. ggsubplot
objects have similar, but different print and build methods than ggplot2
objects.
}

