% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_subplot.r
\name{globalize}
\alias{globalize}
\title{rename global x and y variables in capitals}
\usage{
globalize(obj)
}
\arguments{
\item{obj}{a data.frame}
}
\description{
rename global x and y variables in capitals
}
\keyword{internal}

