% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_layer-class.r
\name{is.sp_layer}
\alias{is.sp_layer}
\title{Is an object (functionally) a sp_layer?}
\usage{
is.sp_layer(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical
}
\description{
Tests whether an object is or ever was a sp_layer.
}

