% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nasa}
\alias{nasa}
\title{Nasa data from the 2006 American Statistical Association Data Expo}
\description{
Nasa data from the 2006 American Statistical Association Data Expo
}
\keyword{data}

