% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{null_omit}
\alias{null_omit}
\title{Remove NULL objects from a list}
\usage{
null_omit(lst)
}
\arguments{
\item{lst}{a list}
}
\description{
null.omit removes the NULL elements from a list and returns the remaining
objects as a more concise list.
}
\keyword{internal}

