% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggsubplot_build.r
\name{which_untrained}
\alias{which_untrained}
\title{find scales that have no range yet}
\usage{
which_untrained(scales)
}
\arguments{
\item{scales}{A list of ggplot2 scales}
}
\description{
find scales that have no range yet
}

