% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_layer_build.r
\name{which_x}
\alias{which_x}
\title{find x scale}
\usage{
which_x(scales)
}
\arguments{
\item{scales}{A list of ggplot2 scales}
}
\description{
which_x picks out the scale that controls x from a list of scales
}

