% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_layer_build.r
\name{which_y}
\alias{which_y}
\title{find y scale}
\usage{
which_y(scales)
}
\arguments{
\item{scales}{A list of ggplot2 scales}
}
\description{
which_y picks out the scale that controls y from a list of scales
}

