% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_coxcomb.r
\name{zero_wrap}
\alias{zero_wrap}
\title{Changes 2 pi to 0}
\usage{
zero_wrap(vec)
}
\arguments{
\item{vec}{a vector of angles in radians
\code{\link[ggplot2]{aes}}}
}
\description{
Changes 2 pi to 0
}
\keyword{internal}

