\docType{data}
\name{casualties}
\alias{casualties}
\title{casualty data from the Afghan War Diary, lightly cleaned}
\description{
  The Afghan War Diary is distributed by the WikiLeaks
  organization and available for download at
  http://www.wikileaks.org/wiki/Afghan_War_Diary,_2004-2010.
  Each row of the data set corresponds to a single casualty
  (one person being injured or killed).
}
\keyword{data}
\keyword{datasets}

