% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aesply.r
\name{condense}
\alias{condense}
\title{reduce a single valued vector to a single element}
\usage{
condense(var)
}
\arguments{
\item{var}{a vector}
}
\value{
a vector of length one if var only contains one unique value, var
otherwise
}
\description{
reduce a single valued vector to a single element
}
\keyword{internal}

