% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_coxcomb.r
\name{coxcomb_sections}
\alias{coxcomb_sections}
\title{coxcomb_sections ensures that continuous fill, alpha, and colour variables
generate groups at build time as if they were discrete}
\usage{
coxcomb_sections(mapping)
}
\arguments{
\item{mapping}{an aesthetic mapping, usually created with
\code{\link[ggplot2]{aes}}}
}
\description{
coxcomb_sections ensures that continuous fill, alpha, and colour variables
generate groups at build time as if they were discrete
}
\keyword{internal}

