% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_layer_build.r
\name{names_scales}
\alias{names_scales}
\title{Returns the first aes of a scale, to use as an identifier for the scale}
\usage{
names_scales(scales)
}
\arguments{
\item{scales}{a list of ggplot2 scales}
}
\description{
Returns the first aes of a scale, to use as an identifier for the scale
}

