% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nasa}
\alias{nasa}
\title{Nasa data from the 2006 American Statistical Association Data Expo}
\format{\preformatted{'data.frame':	41472 obs. of  17 variables:
 $ time       : int  1 1 1 1 1 1 1 1 1 1 ...
 $ y          : int  1 1 1 1 1 1 1 1 1 1 ...
 $ x          : int  1 2 3 4 5 6 7 8 9 10 ...
 $ lat        : num  -21.2 -21.2 -21.2 -21.2 -21.2 -21.2 -21.2 -21.2 -21.2 -21.2 ...
 $ long       : num  -114 -111 -109 -106 -104 ...
 $ date       : POSIXct, format: "1995-01-01" "1995-01-01" ...
 $ cloudhigh  : num  0.5 1.5 1.5 1 0.5 0 0 0 0.5 1 ...
 $ cloudlow   : num  31 31.5 32.5 39 48 50 51 52.5 54 56 ...
 $ cloudmid   : num  2 2.5 3.5 4 4.5 2.5 4.5 5 8.5 11.5 ...
 $ ozone      : int  260 260 260 258 258 258 256 258 256 258 ...
 $ pressure   : int  1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
 $ surftemp   : num  297 297 297 297 296 ...
 $ temperature: num  297 296 296 296 296 ...
 $ id         : chr  "1-1" "2-1" "3-1" "4-1" ...
 $ day        : num  0 0 0 0 0 0 0 0 0 0 ...
 $ month      : num  1 1 1 1 1 1 1 1 1 1 ...
 $ year       : num  1995 1995 1995 1995 1995 ...
}}
\usage{
nasa
}
\description{
Nasa data from the 2006 American Statistical Association Data Expo
}
\keyword{datasets}

