\name{sp_layer-class}
\alias{"[,sp_layer-method"}
\alias{"[<-,sp_layer-method"}
\alias{"+,ggplot,sp_layer-method"}
\alias{"+,ggsubplot,sp_layer-method"}
\alias{"$,sp_layer-method"}
\alias{"$<-,sp_layer-method"}
\alias{"c,sp_layer-method"}
\alias{"ggtransform,sp_layer-method"}
\alias{"ls,sp_layer-method"}
\alias{"rep,sp_layer-method"}
\alias{"show,sp_layer-method"}
\alias{sp_layer-class}
\title{sp_layer class}
\description{
  sp_layers are layers made with ggsubplot methods. They
  are equivalent to the layers made by ggplot2 functions in
  all ways except that they contain extra information that
  is used to divide the data into subplots and locate those
  subplots witihn the layer when plotting.
}

