% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_layer-class.r
\docType{class}
\name{sp_layer-class}
\alias{show,sp_layer-method}
\alias{sp_layer-class}
\title{sp_layer class}
\description{
sp_layers are layers made with ggsubplot methods. They are equivalent to the
layers made by ggplot2 functions in all ways except that they contain extra
information that is used to divide the data into subplots and locate those
subplots witihn the layer when plotting.
}

