% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_subplot.r
\name{vet}
\alias{vet}
\title{Ensure that an object is numeric}
\usage{
vet(x)
}
\arguments{
\item{x}{an R object}
}
\value{
a numeric object
}
\description{
vet tests whether an object is a factor or character string. If so it
attempts to coerce the variable to numeric.
}
\keyword{internal}

